Tests requirng Java 1.8 must be inside usesjava8 package.
