+++ .travis.yml
@@ -7,6 +7,10 @@ jdk:
   - openjdk7
   - openjdk6
 
+install:
+  - chmod -R 777 ./travis/install.sh
+  - ./travis/install.sh
+
 after_success:
   - chmod -R 777 ./travis/after_success.sh
   - ./travis/after_success.sh
