+++ src/test/java/org/apache/ibatis/submitted/usesjava8/use_actual_param_name/User.java
@@ -13,22 +13,26 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.use_actual_param_name;
+package org.apache.ibatis.submitted.usesjava8.use_actual_param_name;
 
-import java.util.List;
+public class User {
 
-import org.apache.ibatis.annotations.Select;
+  private Integer id;
+  private String name;
 
-public interface Mapper {
+  public Integer getId() {
+    return id;
+  }
 
-  @Select("select * from users where id = #{foo}")
-  User getUserById(Integer id);
+  public void setId(Integer id) {
+    this.id = id;
+  }
 
-  @Select("select * from users where id = #{id} and name = #{name}")
-  User getUserByIdAndName(Integer id, String name);
+  public String getName() {
+    return name;
+  }
 
-  List<User> getUsersByIdList(List<Integer> ids);
-
-  List<User> getUsersByIdListAndName(List<Integer> ids, String name);
-
+  public void setName(String name) {
+    this.name = name;
+  }
 }
